CSS, JavaScript & structural upgrades
So, I removed the CSS from the header and placed the calls in
youare/functions.php. I combined print.css into style.scss and compiled
it.

A bunch of refactoring work to make the theme look and work better. The
goal with this commit is to remove a lot of the cruft in the theme and
get it to work well. Still working on the navigation bar, not entirely
sure how I will get THAT working yet. ;)