Removed the Sidebar in single.php and page.php, added youare_archive_x_posts() and youare_posted_date() function.
Okay, so I removed the sidebar on both single.php and page.php. I
figure I will come back and add a feature to re-add it later.

Now, I added youare_archive_x_posts() to the functions file to clean up
the archive.php file. I also changed the date function in single.php to
another new function, youare_posted_date().

If the number of posts you have in an archive view is higher than the
amount of posts per page, it will show the amount of posts per page.
This is on the « to be fixed » list.